% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{Marboha}
\alias{Marboha}
\title{Marboha palette generator}
\usage{
Marboha(n)
}
\arguments{
\item{n}{Number of colors desired.}
}
\value{
A vector of colors.
}
\description{
Marboha palette generator
}
\examples{
Marboha(3)
}
