% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Marboha_palettes}
\alias{Marboha_palettes}
\title{Marboha_palettes palette}
\format{
An object of class \code{list} of length 1.
}
\usage{
Marboha_palettes
}
\description{
Marboha_palettes palette
}
\keyword{datasets}
